Remove hk.jit from the public API.
It is not often that we remove a symbol, but `hk.jit` in its current form is
both not very useful (you can only make it cache **inside** a transformed
function) and a footgun (if you hack up a solution to reuse jit across transform
boundaries you get different caching behaviour in `init` vs. `apply`).

As such we've taken the decision to remove the feature for now while we figure
out a better way to support it.

PiperOrigin-RevId: 432966342